+++ src/main/java/Coordinate/coordikittyBE/domain/security/jwt/JwtAuthenticationFilter.java
@@ -1,13 +1,23 @@
 package Coordinate.coordikittyBE.domain.security.jwt;
 
 import java.io.IOException;
+import java.util.Date;
+import java.util.UUID;
+
+import Coordinate.coordikittyBE.exception.CoordikittyException;
+import Coordinate.coordikittyBE.exception.ErrorType;
+import Coordinate.coordikittyBE.global.util.JwtHelper;
+import io.jsonwebtoken.Claims;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.stereotype.Component;
 import org.springframework.util.StringUtils;
 import org.springframework.web.filter.OncePerRequestFilter;
@@ -15,14 +25,14 @@ import org.springframework.web.filter.OncePerRequestFilter;
 @Component
 @RequiredArgsConstructor
 public class JwtAuthenticationFilter extends OncePerRequestFilter {
-    private final JwtTokenProvider jwtTokenProvider;
-
+    private final JwtHelper jwtHelper;
+    private final UserDetailsService userDetailService;
     @Override
     public void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain chain) throws IOException, ServletException {
         String token = resolveToken(request);
 
-        if (token!=null && jwtTokenProvider.validateToken(token)) {
-            Authentication authentication = jwtTokenProvider.getAuthentication(token);
+        if (token!=null && validateToken(token)) {
+            Authentication authentication = getAuthentication(token);
             SecurityContextHolder.getContext().setAuthentication(authentication);
         }
         chain.doFilter(request, response);
@@ -34,6 +44,29 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
             return bearerToken.substring(7);
         }
         return null;
+    }
+
+    public boolean validateToken(String token) {
+        try {
+            Claims claims = jwtHelper.parseClaims(token);
+            if (claims==null) {
+                throw new CoordikittyException(ErrorType.INVALID_TOKEN);
+            }
+            if(claims.getExpiration().before(new Date())){
+                throw new CoordikittyException(ErrorType.INVALID_TOKEN);
+            }
+            return true;
+        } catch (Exception e) {
+            return false;
+        }
+    }
+
+    public Authentication getAuthentication(String accessToken) {
+        Claims claims = jwtHelper.parseClaims(accessToken);
+        String email = claims.getSubject();
+
+        UserDetails userDetails = userDetailService.loadUserByUsername(email);
+        return new UsernamePasswordAuthenticationToken(userDetails, accessToken, userDetails.getAuthorities());
     }
 }
