+++ src/main/java/Coordinate/coordikittyBE/global/common/jwt/RefreshTokenProvider.java
@@ -19,7 +19,7 @@ public class RefreshTokenProvider implements JwtTokenProvider{
     private final String issuer;
 
     public RefreshTokenProvider(
-            @Value("${jwt.secret}") String secretKey,
+            @Value("${jwt.secret_key}") String secretKey,
             @Value("${jwt.access_expiration}") Duration accessTokenExpiration,
             @Value("${jwt.issuer}") String issuer
     ){
